+++ src/actions/ripple/SendFormActions.js
@@ -457,7 +457,7 @@ export const onSend = (): AsyncAction => async (
         dispatch({
             type: NOTIFICATION.ADD,
             payload: {
-                type: 'error',
+                variant: 'error',
                 title: <FormattedMessage {...l10nMessages.TR_TRANSACTION_ERROR} />,
                 message: signedTransaction.payload.error,
                 cancelable: true,
@@ -476,7 +476,7 @@ export const onSend = (): AsyncAction => async (
         dispatch({
             type: NOTIFICATION.ADD,
             payload: {
-                type: 'error',
+                variant: 'error',
                 title: <FormattedMessage {...l10nMessages.TR_TRANSACTION_ERROR} />,
                 message: push.payload.error,
                 cancelable: true,
@@ -499,7 +499,7 @@ export const onSend = (): AsyncAction => async (
     dispatch({
         type: NOTIFICATION.ADD,
         payload: {
-            type: 'success',
+            variant: 'success',
             title: <FormattedMessage {...l10nMessages.TR_TRANSACTION_SUCCESS} />,
             message: txid,
             cancelable: true,