+++ src/components/modals/passphrase/Passphrase/index.messages.js
@@ -24,6 +24,10 @@ const definedMessages: Messages = defineMessages({
         id: 'TR_PASSPHRASES_DO_NOT_MATCH',
         defaultMessage: 'Passphrases do not match!',
     },
+    TR_PASSPHRASE_IS_TOO_LONG: {
+        id: 'TR_PASSPHRASE_IS_TOO_LONG',
+        defaultMessage: 'Passphrase is too long!',
+    },
     TR_SHOW_PASSPHRASE: {
         id: 'TR_SHOW_PASSPHRASE',
         defaultMessage: 'Show passphrase',