+++ src/views/Wallet/views/Account/Summary/common.messages.js
@@ -23,6 +23,10 @@ const definedMessages: Messages = defineMessages({
         id: 'TR_TOKEN_NOT_FOUND',
         defaultMessage: 'Token not found',
     },
+    TR_ALREADY_ADDED: {
+        id: 'TR_ALREADY_ADDED',
+        defaultMessage: 'Already added',
+    },
     TR_HISTORY: {
         id: 'TR_HISTORY',
         defaultMessage: 'History',